Took world and player structs from `index` and moved them into sepparate files
