+++ src/lib.rs
@@ -2,3 +2,5 @@ pub mod renderer;
 pub mod index;
 pub mod system;
 pub mod manufacture;
+pub mod world;
+pub mod player;
