+++ src/player.rs
@@ -1,0 +1,87 @@
+use crossterm::style::Color;
+
+use crate::*;
+
+/// # Player struct
+/// Use 1-4 in `fp_playerNum` when initializing to change the color
+/// 
+/// Available colors:
+/// * 1 - Cyan
+/// * 2 - Green
+/// * 3 - Yellow
+/// * 4 - Orange
+/// 
+/// # Custom colors
+/// To instead use custom colors set `fp_playerNum` to 0 and `fp_color` to [`Color::Rgb`]
+pub struct TEMPLATE_player {
+    pub p_x: usize,
+    pub p_y: usize,
+    pub p_chunkX: usize,
+    pub p_chunkY: usize,
+    pub p_colorChar: Color,
+    pub p_colorBg: Color,
+}
+impl TEMPLATE_player {
+    pub fn new(INp_playerNum: usize, INp_color: Option<Color>) -> Self{
+        let Fp_playerColor: Color = if INp_playerNum == 0{
+            INp_color.unwrap()
+        }
+        else {
+            GAME_playerColors[INp_playerNum]
+        };
+        TEMPLATE_player {
+            p_x: 11,
+            p_y: 55,
+            p_chunkX: 0,
+            p_chunkY: 0,
+            p_colorChar: Color::White,
+            p_colorBg: Fp_playerColor }
+    }
+    pub fn p_move(&mut self, dir: u8) {
+        match dir {
+            0 => {
+                // Up
+                if self.p_y == 0 {
+                    return;
+                }
+                self.p_y -= 1
+            }
+            1 => {
+                //Down
+                if self.p_y == (system::SYS_GRID_Y - 1) {
+                    return;
+                }
+                self.p_y += 1
+            }
+            2 => {
+                //Left
+                if self.p_x == 0 {
+                    return;
+                }
+                self.p_x -= 1
+            }
+            3 => {
+                //Right
+                if self.p_x == (system::SYS_GRID_X - 1) {
+                    return;
+                }
+                self.p_x += 1
+            }
+            _ => {}
+        }
+    }
+    pub fn p_updateChunkPos(&mut self){
+        self.p_chunkX = (self.p_x / system::SYS_CHUNK_X);
+        self.p_chunkY = (self.p_y / system::SYS_CHUNK_Y);
+    }
+}
+
+/// # Player color "enum"
+/// ## Disclaimer:
+/// Is only for Player 1-4 colors
+pub const GAME_playerColors: [Color;4] = [
+    Color::Cyan,
+    Color::Green,
+    Color::Yellow,
+    Color::Rgb {r: 255, g: 153, b: 0}
+];
